+++ importers/src/Service/HasCoverService.php
@@ -15,6 +15,7 @@ class HasCoverService
 {
     /**
      * @param HttpClientInterface $client
+     * @param bool $enabled
      * @param string $hasCoverServiceUrl
      * @param MetricsService $metricsService
      * @param AuthenticationService $authenticationService
@@ -41,7 +42,7 @@ class HasCoverService
      */
     public function post(string $pid, bool $coverExists): void
     {
-        // Do not send request if service is enabled.
+        // Do not send request if service is not enabled.
         if (!$this->enabled) {
             return;
         }