+++ importers/src/Service/VendorService/BogPortalen/BogPortalenVendorService.php
@@ -120,10 +120,10 @@ class BogPortalenVendorService implements VendorServiceImporterInterface, Vendor
     /**
      * {@inheritDoc}
      */
-    public function getUnverifiedVendorImageItem(string $identifier, string $type): ?UnverifiedVendorImageItem
+    public function getUnverifiedVendorImageItems(string $identifier, string $type): \Generator
     {
-        if (!$this->supportsIdentifierType($type)) {
-            throw new UnsupportedIdentifierTypeException('Unsupported single identifier type: '.$type);
+        if (!$this->supportsIdentifier($identifier, $type)) {
+            throw new UnsupportedIdentifierTypeException(\sprintf('Unsupported single identifier: %s (%s)', $identifier, $type));
         }
 
         $vendor = $this->vendorCoreService->getVendor(self::VENDOR_ID);
@@ -132,13 +132,13 @@ class BogPortalenVendorService implements VendorServiceImporterInterface, Vendor
         $item->setIdentifier($identifier);
         $item->setIdentifierType($type);
 
-        return $item;
+        yield $item;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function supportsIdentifierType(string $type): bool
+    public function supportsIdentifier(string $identifier, string $type): bool
     {
         return IdentifierType::ISBN === $type;
     }