+++ importers/src/Service/VendorService/EbookCentral/EbookCentralVendorService.php
@@ -38,10 +38,10 @@ class EbookCentralVendorService extends AbstractDataWellVendorService implements
     /**
      * {@inheritDoc}
      */
-    public function getUnverifiedVendorImageItem(string $identifier, string $type): ?UnverifiedVendorImageItem
+    public function getUnverifiedVendorImageItems(string $identifier, string $type): \Generator
     {
-        if (!$this->supportsIdentifierType($type)) {
-            throw new UnsupportedIdentifierTypeException('Unsupported single identifier type: '.$type);
+        if (!$this->supportsIdentifier($identifier, $type)) {
+            throw new UnsupportedIdentifierTypeException(\sprintf('Unsupported single identifier: %s (%s)', $identifier, $type));
         }
 
         if (!$this->tools->isValidIsbn13($identifier)) {
@@ -55,13 +55,13 @@ class EbookCentralVendorService extends AbstractDataWellVendorService implements
         $item->setIdentifier($identifier);
         $item->setIdentifierType($type);
 
-        return $item;
+        yield $item;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function supportsIdentifierType(string $type): bool
+    public function supportsIdentifier(string $identifier, string $type): bool
     {
         return IdentifierType::ISBN === $type;
     }