+++ importers/CHANGELOG.md
@@ -18,9 +18,13 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 ### Changed
 - Simplify default validation to be based only on HTTP response code.
 - Allow for custom validators by vendor. This ensures non-standard behavior is scope to the relevant Vendor
+- Change `SingleIdentifier` vendors supports function to consider both type and identifier
+- Changed `VendorServiceSingleIdentifierInterface` to yield `UnverifiedVendorImageItem` items back  
 
 ### Added
 - Added ETag to check for "image updated"
+- Added support for generic covers
+- Added forsider.dk generic cover vendor
 
 ## [3.5.5] - 2023-02-13